allow using a custom attribute for the lightbox title
Even though [title] is apropriate for most use cases, it would be good
to have some flexibility and allow the use of another attribute for
the lightbox title (e.g. [data-title]). This is for cases in which
the tooltip does not make sense.

This commit introduces the `titleattr` option (which defaults to
'title'), to allow such flexibility.