Update venobox.js
Added an option to include a close message as an option to prevent accidental clicking of the overlay. If they hit escape or click the close button specifically, it will close normally, but if they click the overlay it will confirm if they want to close.

If no close message is set, it won't confirm close. The option is "closemessage"